[refactor] model에서 dto로 변경
