--- data/src/main/java/com/prac/data/repository/dto/TokenDto.kt
@@ -1,6 +1,0 @@
-package com.prac.data.repository.dto
-
-internal data class TokenDto(
-    val accessToken: String,
-    val refreshToken: String
-)
