+++ data/src/main/java/com/prac/data/repository/model/RepoModel.kt
@@ -1,11 +1,11 @@
-package com.prac.data.repository.dto
+package com.prac.data.repository.model
 
 import com.prac.data.di.entity.RepoEntity
 
-internal data class RepoDto(
+internal data class RepoModel(
     val id: Int,
     val name: String,
-    val owner: OwnerDto,
+    val owner: OwnerModel,
     val stargazersCount: Int,
     val updatedAt: String
 ) {
