+++ data/src/main/java/com/prac/data/repository/model/TokenModel.kt
@@ -1,0 +1,6 @@
+package com.prac.data.repository.model
+
+internal data class TokenModel(
+    val accessToken: String,
+    val refreshToken: String
+)
