+++ data/src/main/java/com/prac/data/source/RepoApiDataSource.kt
@@ -1,9 +1,9 @@
 package com.prac.data.source
 
-import com.prac.data.repository.dto.RepoDto
+import com.prac.data.repository.model.RepoModel
 
 internal interface RepoApiDataSource {
     suspend fun getRepos(
         userName: String
-    ) : List<RepoDto>
+    ) : List<RepoModel>
 }
