+++ data/src/main/java/com/prac/data/source/TokenApiDataSource.kt
@@ -1,9 +1,9 @@
 package com.prac.data.source
 
-import com.prac.data.repository.dto.TokenDto
+import com.prac.data.repository.model.TokenModel
 
 internal interface TokenApiDataSource {
     suspend fun getToken(
         code: String
-    ) : TokenDto
+    ) : TokenModel
 }
