+++ data/src/main/java/com/prac/data/source/dto/OwnerDto.kt
@@ -1,6 +1,6 @@
 package com.prac.data.source.dto
 
-import com.prac.data.repository.dto.OwnerDto
+import com.prac.data.repository.model.OwnerModel
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -10,5 +10,5 @@ internal data class OwnerDto(
     @SerialName("avatar_url") val avatarUrl: String
 ) {
     fun toModel() =
-        OwnerDto(login, avatarUrl)
+        OwnerModel(login, avatarUrl)
 }
