+++ data/src/main/java/com/prac/data/source/impl/TokenApiDataSourceImpl.kt
@@ -1,6 +1,6 @@
 package com.prac.data.source.impl
 
-import com.prac.data.repository.dto.TokenDto
+import com.prac.data.repository.model.TokenModel
 import com.prac.data.source.TokenApiDataSource
 import com.prac.data.source.api.GitHubTokenApi
 import javax.inject.Inject
@@ -10,12 +10,12 @@ internal class TokenApiDataSourceImpl @Inject constructor(
 ) : TokenApiDataSource {
     override suspend fun getToken(
         code: String
-    ): TokenDto {
+    ): TokenModel {
         val response = gitHubTokenApi.getAccessTokenApi(
             code = code
         )
 
-        return TokenDto(
+        return TokenModel(
             accessToken = response.accessToken,
             refreshToken = response.refreshToken
         )
